Increase API retries to 6
Motivation for the amount of retries being 6 is to give the function a full minute rest (in the worst case scenario) before a final retry. The total wait time in the worst case scanario would be about 2 minutes which would still allow the Lambda to start another process in case it was close to running out of time, the request succeeded and the file wasn't done processing.